requirements.txt
annotated-types==0.7.0
anyio==4.8.0
attrs==25.1.0
certifi==2025.1.31
charset-normalizer==3.4.1
distro==1.9.0
exceptiongroup==1.2.2
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
inflection==0.5.1
jiter==0.8.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
openai==1.61.0
openapi==2.0.0
pydantic==2.10.6
pydantic_core==2.27.2
pyTelegramBotAPI==4.26.0
python-dotenv==1.0.1
referencing==0.36.2
requests==2.32.3
rpds-py==0.22.3
sniffio==1.3.1
tqdm==4.67.1
typing_extensions==4.12.2
urllib3==2.3.0
